requirements.txt
# 核心依赖
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
joblib==1.3.2
mlflow==3.5.1

# 测试依赖
pytest==8.3.3
pytest-cov==4.1.0

# 代码质量工具
black==23.7.0
flake8==6.0.0
isort==5.12.0
mypy==1.5.1

# 可选：数据可视化
matplotlib==3.7.2
seaborn==0.12.2